Changed tag names for: div class='main-nav' . div class='card'
Text pasted from Slack:
"Changed tag names for the following div elements: div class='main-nav' should be a nav tag. div class='card' should be a section or article tag.
In your index.html, you have missed changing the div tags for the navigation and the div with the H2 heading Goals. If you are truly going for Exceeds, you also need to change the div tags in your other 2 HTML files. You have correctly changed most of the tags in your index.html page. So you could do it for the others.
So you just have above to take care of for exceeds part above.
Can I also recommend that you give your social media links the attribute target with the value "_blank" so when a user clicks on your social media links the links will open in a new tab leaving your website still open in a tab.
You don't have to, this is only a suggestion. "